requirements.txt
falcon==3.0.1
google-cloud-storage
google-cloud-secret-manager
google-auth
python-dotenv
scenedetect[opencv-headless]